refactor: change modifier extensions to new package
